Add debug logging for core intentlayzer module
Started peppering the base intentalyzer module with `debug` logging
statements.

Additional housekeeping:
  * Marked all members of `RecognizedIntent<>` as `readonly` to
"encourage" immutability
  * Refactored the `createRecognizerChain` (all overloads)
and `createEnricherPipeline` (all overloads) into their own files.
  * Refactored implementation of `enrichSpecificIntent`
