+++ packages/intentalyzer/src/index.ts
@@ -1,5 +1,6 @@
 export * from "./core-types";
+export * from "./create-enricher-pipeline";
 export * from "./create-intent-resolver";
+export * from "./create-recognizer-chain";
 export * from "./entities";
 export * from "./utilities";
-
